pre-release branch is verified

+++ user_space/sdcard_boot_update.sh
@@ -1,9 +1,16 @@
 #!/bin/bash
+if [ "$#" -ne 1 ]; then
+    echo "You must enter the \$BOARD_NAME as argument"
+    echo "Like: adrv9364z7020 adrv9361z7035 adrv9361z7035_fmc zc706_fmcs2 zed_fmcs2 zc702_fmcs2"
+    exit 1
+fi
+BOARD_NAME=$1
+
 set -x
 
 rm BOOT.BIN
 sync
-wget ftp://192.168.10.1/kernel_boot/output_boot_bin/BOOT.BIN
+wget ftp://192.168.10.1/kernel_boot/boards/$BOARD_NAME/output_boot_bin/BOOT.BIN
 sync
 rm uImage
 sync
@@ -11,18 +18,18 @@ wget ftp://192.168.10.1/adi-linux/arch/arm/boot/uImage
 sync
 rm devicetree.dtb
 sync
-wget ftp://192.168.10.1/kernel_boot/devicetree.dtb
+wget ftp://192.168.10.1/kernel_boot/boards/$BOARD_NAME/devicetree.dtb
 sync
 #slepp 0.5
-mount /dev/mmcblk0p1  /sdcard
+mount /dev/mmcblk0p1  /mnt
 sync
 #sleep 0.5
-cp BOOT.BIN /sdcard/ -f
-cp uImage /sdcard/ -f
-cp devicetree.dtb /sdcard/ -f
+cp BOOT.BIN /mnt/ -f
+cp uImage /mnt/ -f
+cp devicetree.dtb /mnt/ -f
 sync
 #sleep 0.5
-umount /sdcard
+umount /mnt
 sync
 #sleep 3
 sudo reboot now
